+++ plone/recipe/codeanalysis/__init__.py
@@ -1,11 +1,16 @@
 # -*- coding: utf-8 -*-
+from plone.recipe.codeanalysis.chameleonlint import ChameleonLint
 from plone.recipe.codeanalysis.check_manifest import CheckManifest
 from plone.recipe.codeanalysis.clean_lines import CleanLines
 from plone.recipe.codeanalysis.csslint import CSSLint
+from plone.recipe.codeanalysis.dependencychecker import DependencyChecker
 from plone.recipe.codeanalysis.flake8 import Flake8
 from plone.recipe.codeanalysis.i18ndude import I18NDude
+from plone.recipe.codeanalysis.importchecker import ImportChecker
 from plone.recipe.codeanalysis.jscs import JSCS
 from plone.recipe.codeanalysis.jshint import JSHint
+from plone.recipe.codeanalysis.scsslint import SCSSLint
+from plone.recipe.codeanalysis.xmllint import XMLLint
 from plone.recipe.codeanalysis.zptlint import ZPTLint
 from time import time
 
@@ -18,13 +23,18 @@ import zc.recipe.egg
 
 current_dir = os.path.dirname(__file__)
 all_checks = [
+    ChameleonLint,
     CSSLint,
     CheckManifest,
     CleanLines,
+    DependencyChecker,
+    ImportChecker,
     Flake8,
     I18NDude,
     JSCS,
     JSHint,
+    SCSSLint,
+    XMLLint,
     ZPTLint,
 ]
 
@@ -63,6 +73,8 @@ class Recipe(object):
         self.options.setdefault('jscs', 'False')
         self.options.setdefault('jscs-bin', 'jscs')
         self.options.setdefault('jscs-exclude', '')
+        # Chameleon Lint
+        self.options.setdefault('chameleon-lint', 'False')
         # CSS Lint
         self.options.setdefault('csslint', 'False')
         self.options.setdefault('csslint-bin', 'csslint')
@@ -72,6 +84,12 @@ class Recipe(object):
         # clean lines
         self.options.setdefault('clean-lines', 'False')
         self.options.setdefault('clean-lines-exclude', '')
+        # dependencychecker
+        self.options.setdefault('dependencychecker', 'False')
+        self.options.setdefault('dependencychecker-bin', 'dependencychecker')
+        # importchecker
+        self.options.setdefault('importchecker', 'False')
+        self.options.setdefault('importchecker-bin', 'importchecker')
         # Jenkins output
         self.options.setdefault('jenkins', 'False')
         self.options.setdefault('flake8-filesystem', 'False')
@@ -80,6 +98,12 @@ class Recipe(object):
         # Find untranslated strings
         self.options.setdefault('find-untranslated', 'False')
         self.options.setdefault('i18ndude-bin', '')
+        # scss-lint
+        self.options.setdefault('scsslint', 'False')
+        self.options.setdefault('scsslint-bin', 'scss-lint')
+        self.options.setdefault('scsslint-config', '')
+        # xmllint
+        self.options.setdefault('xmllint', 'False')
         # zptlint
         self.options.setdefault('zptlint', 'False')
         self.options.setdefault('zptlint-bin', '')