+++ plone/recipe/codeanalysis/i18ndude.py
@@ -7,16 +7,19 @@ class I18NDude(Analyser):
 
     name = 'find-untranslated'
     title = 'Translations'
+    extensions = ('pt', 'cpt', 'zpt', )
 
     @property
     def cmd(self):
         cmd = []
-        files = self.find_files('.*\.pt')
+        files = []
+        for extension in self.extensions:
+            files.extend(self.find_files('.*\.{0}'.format(extension)))
 
         if files:
             cmd.append(self.options.get('i18ndude-bin') or '')
             cmd.append('find-untranslated')
-            if self.nosummary:
+            if self.nosummary or self.use_jenkins:
                 cmd.append('--nosummary')
             cmd.extend(files)