+++ plone/recipe/codeanalysis/jscs.py
@@ -19,7 +19,7 @@ class JSCS(Analyser):
         cmd = []
         files = self.find_files('.*\.js')
 
-        if files:
+        if files or self.use_jenkins:
             cmd.append(self.get_prefixed_option('bin'))
 
             if self.use_jenkins: