+++ plone/recipe/codeanalysis/tests/test_docs.py
@@ -5,7 +5,6 @@ import doctest
 import os
 import unittest
 import zc.buildout.testing
-import zc.buildout.tests
 
 
 optionflags = (
@@ -28,6 +27,7 @@ def setUp(test):  # noqa
 
     # Install second level install_requires dependencies
     zc.buildout.testing.install('pep8', test)
+    zc.buildout.testing.install('lxml', test)
     zc.buildout.testing.install('pyflakes', test)
     zc.buildout.testing.install('mccabe', test)
     zc.buildout.testing.install('zope.exceptions', test)